requirements.txt
numpy
torch-cluster
torch-geometric
torch-scatter
torch-sparse
torch-spline-conv
tensorboard
tqdm
networkx
cnfgen
pycsp3
